+++ frontend/src/scenes/dashboard/DashboardHeader.js
@@ -7,14 +7,12 @@ import React from 'react'
 import { useActions, useValues } from 'kea'
 import { dashboardsModel } from '~/models/dashboardsModel'
 import { PushpinFilled, PushpinOutlined, EllipsisOutlined, EditOutlined, DeleteOutlined } from '@ant-design/icons'
-import { newDashboardLogic } from './newDashboardLogic'
 
-export function DashboardHeader({ logic, id }) {
+export function DashboardHeader({ logic }) {
     const { dashboard } = useValues(logic)
-    const { renameDashboard } = useActions(logic)
+    const { addNewDashboard, renameDashboard } = useActions(logic)
     const { dashboards, dashboardsLoading } = useValues(dashboardsModel)
     const { pinDashboard, unpinDashboard, deleteDashboard } = useActions(dashboardsModel)
-    const { addNewDashboard } = useActions(newDashboardLogic({ key: `dashboard-${id}`, redirect: true }))
 
     return (
         <div className="dashboard-header">
